requirements.txt
requests
pillow
python-dotenv
sqlalchemy
pymysql
